winegard_pathway_x2_rotor_control/requirements.txt
#Package requirements for winegard_pathway_x2_rotor_control.py

pyserial
regex
socket
